+++ Magic/CHANGELOG.md
@@ -8,6 +8,7 @@
  - Fix mana not saving properly on droppable wands
  - Fix Colloportus in 1.14+
  - Add require_mana_percentage parameter to CheckMana action
+ - Add item_count parameter to GiveItem action so it can be used with variables
 
 # 7.9.6